% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calib.R
\name{calib}
\alias{calib}
\title{The function calib}
\usage{
calib(conc, OD, ...)
}
\arguments{
\item{conc}{description of parameter}

\item{OD}{takes the optical densities}

\item{...}{currently disregarded}
}
\value{
the lm information
}
\description{
It calculates a new cocnentration based on a cailbration curve,
}
\details{
here you can give more details

Those detail can be multiple paragraphs
}
\examples{
conc <- seq(1:10)
OD <- seq(1:10) + rnorm (10)
calib(conc, OD)

}
